requirements.txt
aiogram==3.0.0b7
numpy==1.25.1
Pillow==10.0.0
Pillow==10.0.0
torch==2.0.1
gunicorn=20.1.0
